+++ service/api/endpoints/modeling.py
@@ -1,35 +1,28 @@
-from typing import Any, Dict, List, Sequence, Union
+from typing import Any, Dict, List
 
 import numpy as np
 from aiobotocore.session import ClientCreatorContext
 from bertopic import BERTopic
-from fastapi import Depends, Path
+from fastapi import Depends
 from fastapi.exceptions import HTTPException
-from fastapi.responses import JSONResponse
 from fastapi.routing import APIRouter
-from fastapi_pagination import LimitOffsetPage
-from fastapi_pagination.bases import AbstractPage
-from pydantic.types import UUID4
-from sqlalchemy.exc import NoResultFound
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from ... import crud
-from ...api import deps
-from ...api.utils import get_sample_dataset, load_model, save_model
-from ...core.config import settings
 from ...models import models
 from ...schemas.base import (
     DocsWithPredictions,
     FitResult,
     Input,
-    Message,
     ModelId,
     ModelPrediction,
     PredictIn,
 )
 from ...schemas.bertopic_wrapper import BERTopicWrapper
+from .. import deps
+from ..utils import get_sample_dataset, load_model, save_model
 
-router = APIRouter(tags=["model_training"])
+router = APIRouter(prefix="/modeling", tags=["modeling"])
 
 
 async def gather_topics(topic_model: BERTopic) -> List[Dict[str, Any]]:
@@ -47,7 +40,7 @@ async def gather_topics(topic_model: BERTopic) -> List[Dict[str, Any]]:
     return topics
 
 
-@router.post("/models", summary="Run topic modeling", response_model=FitResult)
+@router.post("/training", summary="Run topic modeling", response_model=FitResult)
 async def fit(
     data: Input,
     s3: ClientCreatorContext = Depends(deps.get_s3),
@@ -76,7 +69,7 @@ async def fit(
 
 
 @router.post(
-    "/models/{model_id}/predict",
+    "/{model_id}/predicting",
     summary="Predict with existing model",
     response_model=ModelPrediction,
 )
@@ -93,7 +86,7 @@ async def predict(
 
 
 @router.post(
-    "/models/{model_id}/reduce-topics",
+    "/{model_id}/reducting",
     summary="Reduce number of topics in existing model",
     response_model=FitResult,
 )
@@ -135,60 +128,3 @@ async def reduce_topics(
         ),
         predictions=ModelPrediction(topics=predicted_topics, probabilities=probs.tolist()),
     )
-
-
-@router.get(
-    "/models", summary="Get existing models", response_model=LimitOffsetPage[models.TopicModelBase]
-)
-async def list_models(
-    session: AsyncSession = Depends(deps.get_db_async),
-) -> AbstractPage[models.TopicModel]:
-    return await crud.topic_model.paginate(session)
-
-
-@router.get("/models/{model_id}/", summary="Get topics", response_model=List[models.TopicBase])
-async def get_topics(
-    model_id: UUID4 = Path(...),
-    version: int = 1,
-    session: AsyncSession = Depends(deps.get_db_async),
-) -> Sequence[models.Topic]:
-    result: Sequence[models.Topic] = await crud.topic.get_model_topics(
-        session, model_id=model_id, version=version
-    )
-    return result
-
-
-@router.get(
-    "/models/{model_id}/topics",
-    summary="Get topics with words",
-    response_model=List[models.TopicWithWords],
-)
-async def get_topics_info(
-    model_id: UUID4 = Path(...),
-    version: int = 1,
-    session: AsyncSession = Depends(deps.get_db_async),
-) -> Sequence[models.TopicWithWords]:
-    result: Sequence[models.TopicWithWords] = await crud.topic.get_model_topics(
-        session, model_id=model_id, version=version, with_words=True
-    )
-    return result
-
-
-@router.delete(
-    "/models/{model_id}",
-    summary="Remove topic model",
-    responses={404: {"model": Message}},
-    response_model=Message,
-)
-async def remove_model(
-    model_id: UUID4 = Path(...),
-    version: int = 1,
-    s3: ClientCreatorContext = Depends(deps.get_s3),
-    session: AsyncSession = Depends(deps.get_db_async),
-) -> Union[Message, JSONResponse]:
-    try:
-        await crud.topic_model.remove_by_id_version(session, model_id=model_id, version=version)
-        await s3.delete_object(Bucket=settings.MINIO_BUCKET_NAME, Key=str(model_id))
-    except (NoResultFound, s3.exceptions.NoSuchKey):
-        return JSONResponse(status_code=404, content=dict(Message(message="Model not found")))
-    return Message(message="ok")
