+++ tests/test_modeling.py
@@ -4,12 +4,13 @@ pytestmark = pytest.mark.slow
 
 
 def test_fit(client):
-    response = client.post("/models", json={"texts": []})
+    response = client.post("/modeling/training", json={"texts": []})
     assert response.status_code == 200
 
     response_json = response.json()
-    assert "model_id" in response_json
-    assert "version" in response_json and response_json["version"] == 1
+    assert "model" in response_json
+    assert "model_id" in response_json["model"]
+    assert "version" in response_json["model"] and response_json["model"]["version"] == 1
     assert "predictions" in response_json
     assert len(response_json["predictions"]["topics"]) == len(
         response_json["predictions"]["probabilities"]
